Moved handlers to their respective places.
